requirements.txt
aiohttp==3.10.6
pandas==2.2.2
pillow==10.4.0
prometheus-api-client==0.5.5
requests==2.32.3
slack_bolt==1.20.1
XlsxWriter==3.2.0
python-dateutil==2.9.0.post0
redis==5.0.8
rq==1.16.2
selenium==4.25.0
PyYAML==6.0.2